BestTime updates `PoppinLocation.ForecastWeek` any time a location is retrieved if it's missing or more than 14 days old It also stores a new forecast when a `PoppinLocation` is created.

+++ Models/BusinessEntities/PoppinLocationEntities.cs
@@ -159,6 +159,6 @@ namespace Poppin.Models.BusinessEntities
 
 	public interface IForecast
 	{
-		public double GetForecastOccupancy();
+		public int GetForecastOccupancy();
 	}
 }
